feat: Add ash tree question to Step6 component
This commit adds a new question to the Step6 component asking users if the sighting is on an ash tree. The question is implemented using radio buttons with three options: "Yes" true, "No" false , and "Not Sure" remains undefined. The `handleAshTreeChange` function is used to update the `formData` state with the selected option.