feat: Add comments and specimen checkbox to Step9 component
This commit adds a new textarea input to the Step9 component for users to leave comments about their sighting. The `handleChange` function is used to update the `formData` state with the entered comments.

This commit also adds a new checkbox to the Step9 component asking users if they have a specimen. The checkbox is implemented using the `formData` state and the `handleCheckboxChange` function.