feat: Add image upload functionality to Step10 component
This commit adds image upload functionality to the Step10 component using the `handleImageUpload` and `handleImageChange` functions. The `handleImageUpload` function is used to get the selected image files from the user and pass them to the `handleImageChange` function. The `handleImageChange` function is used to update the `formData` state with the selected images.

The selected images are displayed as thumbnails using the `selectedImages` state and the `map` function. The `input` element is used to allow users to select multiple image files.